+++ src/globus_sdk/__init__.py
@@ -21,10 +21,12 @@ from globus_sdk.exc import (
     GlobusError,
     GlobusSDKUsageError,
     GlobusTimeoutError,
+    GroupsAPIError,
     NetworkError,
     SearchAPIError,
     TransferAPIError,
 )
+from globus_sdk.groups import GroupsClient
 from globus_sdk.local_endpoint import LocalGlobusConnectPersonal
 from globus_sdk.response import GlobusHTTPResponse, GlobusResponse
 from globus_sdk.search import SearchClient, SearchQuery
@@ -42,6 +44,7 @@ __all__ = (
     "AuthAPIError",
     "TransferAPIError",
     "SearchAPIError",
+    "GroupsAPIError",
     "NetworkError",
     "GlobusConnectionError",
     "GlobusTimeoutError",
@@ -60,6 +63,7 @@ __all__ = (
     "DeleteData",
     "SearchClient",
     "SearchQuery",
+    "GroupsClient",
     "LocalGlobusConnectPersonal",
 )