+++ src/globus_sdk/groups/__init__.py
@@ -1,0 +1,3 @@
+from .client import GroupsClient
+
+__all__ = ("GroupsClient",)